Related work items: #426, #574, #575, #577, #579, #580, #581, #582, #583, #585, #586, #589

+++ Webzine.WebApplication/Program.cs
@@ -1,51 +1,79 @@
 using Microsoft.EntityFrameworkCore;
 using NLog.Web;
+using Webzine.Business;
+using Webzine.Business.Contracts;
 using Webzine.EntitiesContext;
+using Webzine.EntitiesContext.Seeders;
 using Webzine.Repository;
 using Webzine.Repository.Contracts;
+using Webzine.WebApplication.Filters;
+using Webzine.WebApplication.Middlewares;
 
+// Créer un nouveau builder pour l'application web.
 var builder = WebApplication.CreateBuilder(args);
 
-// Vérifie le type de SGBD à utiliser en fonction de la configuration.
-if (builder.Configuration.GetSection("AppSettings:SGBD").Value == "SQLite")
-{
-    // Utilise SQLite comme SGBD.
-    builder.Services.AddDbContext<WebzineDbContext>(options =>
-  options.UseSqlite(builder.Configuration.GetConnectionString("SqliteConnection")));
-}
-else if (builder.Configuration.GetSection("AppSettings:SGBD").Value == "PostgreSQL")
-{
-    // Utilise PostgreSQL comme SGBD.
-    builder.Services.AddDbContext<WebzineDbContext>(options =>
-   options.UseNpgsql(builder.Configuration.GetConnectionString("PostgresConnection")));
-}
+var seederValue = builder.Configuration.GetSection("Seeder").Value;
+var repositoryValue = builder.Configuration.GetSection("Repository").Value;
 
-// Configure le type de repository à utiliser en fonction de la configuration.
-if (builder.Configuration.GetSection("AppSettings:Repository").Value == "Local")
+// Choix de l'utilisateur : Configure le type de Repository à utiliser en fonction de la configuration.
+if (repositoryValue == "db")
 {
-    // Utilise des repositories locaux.
-    builder.Services.AddScoped<IArtisteRepository, LocalArtisteRepository>();
-    builder.Services.AddScoped<ITitreRepository, LocalTitreRepository>();
-    builder.Services.AddScoped<IStyleRepository, LocalStyleRepository>();
-    builder.Services.AddScoped<ICommentaireRepository, LocalCommentaireRepository>();
-}
-else
-{
-    // Utilise des repositories de base de données.
+    if (builder.Environment.IsDevelopment())
+    {
+        // Si environnement de développement : SGBD SQLite utilisé
+        builder.Services.AddDbContext<WebzineDbContext>(
+            options =>
+            options.UseSqlite(builder.Configuration.GetConnectionString("SqliteConnection")),
+            ServiceLifetime.Scoped);
+    }
+    else
+    {
+        // Si environnement de production : SGBD Postgre utilisé
+        builder.Services.AddDbContext<WebzineDbContext>(
+            options =>
+            options.UseNpgsql(builder.Configuration.GetConnectionString("PostgresConnection")),
+            ServiceLifetime.Scoped);
+    }
+
+    // Créer une base de données.
     builder.Services.AddScoped<IArtisteRepository, DbArtisteRepository>();
     builder.Services.AddScoped<ITitreRepository, DbTitreRepository>();
     builder.Services.AddScoped<IStyleRepository, DbStyleRepository>();
     builder.Services.AddScoped<ICommentaireRepository, DbCommentaireRepository>();
 }
+else
+{
+    // Ne pas utiliser de base de données.
+    builder.Services.AddScoped<IArtisteRepository, LocalArtisteRepository>();
+    builder.Services.AddScoped<ITitreRepository, LocalTitreRepository>();
+    builder.Services.AddScoped<IStyleRepository, LocalStyleRepository>();
+    builder.Services.AddScoped<ICommentaireRepository, LocalCommentaireRepository>();
+  
+}
+
+builder.Services.AddScoped<IDashboardService, DashboardService>();
 
 // Configure les services nécessaires.
 builder.Services.AddControllersWithViews()
     .AddRazorRuntimeCompilation();
 
+// Injecter les filtres
+builder.Services.AddControllersWithViews(options =>
+{
+    options.Filters.Add(typeof(LoggingActionFilter));
+});
+
 // Initialise NLog.
 builder.Logging.ClearProviders();
 builder.Host.UseNLog();
 
+// Ajoutez la configuration du système de journalisation ici.
+builder.Services.AddLogging(builder =>
+{
+    builder.AddConsole();
+});
+
+// Construit l'application en utilisant les configurations précédemment définies.
 var app = builder.Build();
 
 // Redirection HTTPS si ce n'est pas en développement.
@@ -53,6 +81,99 @@ if (!app.Environment.IsDevelopment())
 {
     app.UseHttpsRedirection();
 }
+
+// Active la possibilité de servir des fichiers statiques présents dans le dossier wwwroot.
+app.UseStaticFiles();
+
+// Active le middleware permettant le routage des requêtes entrantes.
+app.UseRouting();
+
+// Utilise le middleware personnalisé pour la journalisation des requêtes entrantes.
+app.UseMiddleware<RequestLoggingMiddleware>();
+
+// Générer une page 404
+app.UseStatusCodePagesWithReExecute("/StatusCode/Error{0}");
+
+// Mise en forme des routes
+
+// Routes specifiques de l'administration :
+
+// Liste des artistes
+app.MapControllerRoute(
+    name: "artistesAvecPage",
+    pattern: "administration/artistes/page/{NumeroPage}",
+    defaults: new { area = "Administration", controller = "Artiste", action = "Index" });
+
+app.MapControllerRoute(
+    name: "artistesSansPage",
+    pattern: "administration/artistes/",
+    defaults: new { area = "Administration", controller = "Artiste", action = "Index" });
+
+
+// Liste des titres
+app.MapControllerRoute(
+    name: "titresAvecPage",
+    pattern: "administration/titres/page/{NumeroPage}",
+    defaults: new { area = "Administration", controller = "Titre", action = "Index" });
+
+app.MapControllerRoute(
+    name: "titresSansPage",
+    pattern: "administration/titres/",
+    defaults: new { area = "Administration", controller = "Titre", action = "Index" });
+
+// Liste des styles
+app.MapControllerRoute(
+    name: "stylesAvecPage",
+    pattern: "administration/styles/page/{NumeroPage}",
+    defaults: new { area = "Administration", controller = "Style", action = "Index" });
+
+app.MapControllerRoute(
+    name: "stylesSansPage",
+    pattern: "administration/styles/",
+    defaults: new { area = "Administration", controller = "Style", action = "Index" });
+
+// Routes specifiques de la consultation :
+
+// Liste des commentaires
+app.MapControllerRoute(
+    name: "commentairesAvecPage",
+    pattern: "administration/commentaires/page/{NumeroPage}",
+    defaults: new { area = "Administration", controller = "Commentaire", action = "Index" });
+
+app.MapControllerRoute(
+    name: "commentairesSansPage",
+    pattern: "administration/commentaires/",
+    defaults: new { area = "Administration", controller = "Commentaire", action = "Index" });
+
+// Titres selon le style de musique démandé
+app.MapControllerRoute(
+    name: "style",
+    pattern: "titres/style/{style}/",
+    defaults: new { controller = "Titre", action = "Style" });
+
+// Titre par Id
+app.MapControllerRoute(
+    name: "titre",
+    pattern: "titre/{idTitre}",
+    defaults: new { controller = "Titre", action = "Index" });
+
+// Formulaire pour ajouter un commentaire au titre
+app.MapControllerRoute(
+    name: "commenter",
+    pattern: "commenter/",
+    defaults: new { controller = "Titre", action = "Commenter" });
+
+// Page d'un artiste
+app.MapControllerRoute(
+    name: "artiste",
+    pattern: "artiste/{nom}",
+    defaults: new { controller = "Artiste", action = "Index" });
+
+// Route pour les pages à l'accueil
+app.MapControllerRoute(
+    name: "accueilPage",
+    pattern: "page/{NumeroPage?}",
+    defaults: new { controller = "Home", action = "Index" });
 
 // Définit les routes pour les controllers.
 app.MapControllerRoute(
@@ -63,29 +184,37 @@ app.MapControllerRoute(
     name: "default",
     pattern: "{controller=Home}/{action=Index}/{id?}");
 
-// Utilise un scope pour gérer les services.
-using (var scope = app.Services.CreateScope())
+
+if (repositoryValue == "db")
 {
+    // Si le choix est d'avoir une Base de donnée, la seeder :
+    // Utilise un scope pour gérer les services.
+    using var scope = app.Services.CreateScope();
     var services = scope.ServiceProvider;
 
     // Récupère le DbContext du service scope.
     var context = services.GetRequiredService<WebzineDbContext>();
 
-    // Assure la suppression de la base de données.
-    context.Database.EnsureDeleted();
-
-    // Assure la création de la base de données.
+    if (app.Environment.IsDevelopment())
+    {
+        // Environnement de dev : Suppression de la BDD 
+        context.Database.EnsureDeleted();
+    }
+    
     context.Database.EnsureCreated();
+    
 
-    // Opération de seeding
-    SeedData.Initialize(services, context);
+    if (seederValue == "spotify")
+    {
+        // Si la configuration "Seeder" n'est pas vide, seeder la BDD avec les données de Spotify
+        await SeedDataSpotify.Request(context, builder.Configuration.GetSection("spotify"));
+    }
+    else if (seederValue == "local")
+    {
+        // Sinon seeder la BDD avec des fausses données
+        SeedDataLocal.Initialize(context);
+    }
 }
-
-// Active la possibilité de servir des fichiers statiques présents dans le dossier wwwroot.
-app.UseStaticFiles();
-
-// Active le middleware permettant le routage des requêtes entrantes.
-app.UseRouting();
 
 // Exécute l'application.
 app.Run();
